refactor(Game) : Game 클래스의 UI 기능 분리를 위한 리펙토링
